# Deprecate `shuffle_particles` and other optimizations