--- a/ext/_CUDA.jl
+++ b/ext/_CUDA.jl
@@ -105,8 +105,8 @@ module _3D
     function JustPIC._3D.advection_RK!(
         particles::ParticlesExt{CUDABackend},
         V,
-        grid_vx::NTuple{2,T},
-        grid_vy::NTuple{2,T},
+        grid_vx::NTuple{3,T},
+        grid_vy::NTuple{3,T},
         grid_vz::NTuple{3,T},
         dt,
         α,