--- a/ext/_CUDA.jl
+++ b/ext/_CUDA.jl
@@ -60,14 +60,19 @@ module _2D
         return check_injection(particles)
     end
     function JustPIC._2D.inject_particles!(
-        particles::ParticlesExt{CUDABackend}, args::Vararg{Any,N}
+        particles::ParticlesExt{CUDABackend}, args, fields, grid
     ) where {N}
-        return inject_particles!(particles, args...)
+        return inject_particles!(particles, args, fields, grid)
     end
     function JustPIC._2D.shuffle_particles!(
         particles::ParticlesExt{CUDABackend}, args::Vararg{Any,N}
     ) where {N}
         return shuffle_particles!(particles, args...)
+    end
+    function JustPIC._2D.move_particles!(
+        particles::ParticlesExt{CUDABackend}, grid, args
+    )
+        return shuffle_particles!(particles, grid, args)
     end
     function JustPIC._2D.init_cell_arrays(
         particles::ParticlesExt{CUDABackend}, V::Val{N}
@@ -137,14 +142,19 @@ module _3D
         return check_injection(particles)
     end
     function JustPIC._3D.inject_particles!(
-        particles::ParticlesExt{CUDABackend}, args::Vararg{Any,N}
+        particles::ParticlesExt{CUDABackend}, args, fields, grid
     ) where {N}
-        return inject_particles!(particles, args...)
+        return inject_particles!(particles, args, fields, grid)
     end
     function JustPIC._3D.shuffle_particles!(
         particles::ParticlesExt{CUDABackend}, args::Vararg{Any,N}
     ) where {N}
         return shuffle_particles!(particles, args...)
+    end
+    function JustPIC._3D.move_particles!(
+        particles::ParticlesExt{CUDABackend}, grid, args
+    )
+        return shuffle_particles!(particles, grid, args)
     end
     function JustPIC._3D.init_cell_arrays(
         particles::ParticlesExt{CUDABackend}, V::Val{N}