--- a/scripts/rotating_circle.jl
+++ b/scripts/rotating_circle.jl
@@ -1,10 +1,10 @@
+using JustPIC
+using JustPIC._2D
 # Threads is the default backend, 
 # to run on a CUDA GPU load CUDA.jl (i.e. "using CUDA"), 
 # and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU")
 const backend = CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
 
-using JustPIC
-using JustPIC._2D
 using GLMakie
 
 function expand_range(x::AbstractRange)