Includes several performance optimizations. In particular, deprecates `shuffle_particles!` in favour of `move_particles!`. 

For a 64^3 setup with ~31M particles (Julia 1.10; M2 Pro chip; 4 threads):

`#main` branch
```julia
Section     ncalls     time    %tot     avg     alloc    %tot      avg 
TOTAL             1    2.80s  100.0%   2.80s   61.2KiB  100.0%  61.2KiB
   shuffle        1    1.64s   58.6%   1.64s   34.6KiB   56.5%  34.6KiB
   RK2            1    744ms   26.6%   744ms   7.42KiB   12.1%  7.42KiB
   p2g            1    334ms   11.9%   334ms   6.81KiB   11.1%  6.81KiB
   g2p            1   81.6ms    2.9%  81.6ms   5.75KiB    9.4%  5.75KiB
```

this PR
```julia
Section     ncalls     time    %tot     avg     alloc    %tot      avg
TOTAL         1    1.37s   100.0%   1.37s   33.5KiB  100.0%  33.5KiB
   RK2         1    772ms   56.2%   772ms   7.42KiB   22.2%  7.42KiB
   p2g         1    282ms   20.5%   282ms   6.16KiB   18.4%  6.16KiB
   move        1    254ms   18.5%   254ms   5.89KiB   17.6%  5.89KiB
   g2p         1   65.4ms   4.8%  65.4ms   6.16KiB   18.4%  6.16KiB
   inject      1    556μs   0.0%   556μs   2.72KiB    8.1%  2.72KiB
```